Move PersistenceService into DeviceStateStore directly
