+++ Sources/Internal/ServerSyncProcessHandler.swift
@@ -91,9 +91,9 @@ class ServerSyncProcessHandler {
                     }
                 }
 
-                let localInterestsWillChange = Set(DeviceStateStore.interestsService.getSubscriptions() ?? []) != interestsSet
+                let localInterestsWillChange = Set(DeviceStateStore().getSubscriptions() ?? []) != interestsSet
                 if localInterestsWillChange {
-                    DeviceStateStore.interestsService.persist(interests: Array(interestsSet))
+                    DeviceStateStore().persist(interests: Array(interestsSet))
                     self.handleServerSyncEvent(.InterestsChangedEvent(interests: Array(interestsSet)))
                 }
 
@@ -102,7 +102,7 @@ class ServerSyncProcessHandler {
                 Device.persist(device.id)
             }
 
-            let localInterests = DeviceStateStore.interestsService.getSubscriptions() ?? []
+            let localInterests = DeviceStateStore().getSubscriptions() ?? []
             let remoteInterestsWillChange = Set(localInterests) != device.initialInterestSet ?? Set()
             if remoteInterestsWillChange {
                 // We don't care about the result at this point.
@@ -121,8 +121,8 @@ class ServerSyncProcessHandler {
         Device.delete()
         Device.deleteAPNsToken()
         Metadata.delete()
-        DeviceStateStore.interestsService.persistServerConfirmedInterestsHash("")
-        DeviceStateStore.usersService.removeUserId()
+        DeviceStateStore().persistServerConfirmedInterestsHash("")
+        DeviceStateStore().removeUserId()
         self.handleServerSyncEvent(.StopEvent)
     }
 
@@ -135,13 +135,13 @@ class ServerSyncProcessHandler {
             }
         }
 
-        let localInterests = DeviceStateStore.interestsService.getSubscriptions() ?? []
+        let localInterests = DeviceStateStore().getSubscriptions() ?? []
         let localInterestsHash = localInterests.calculateMD5Hash()
 
-        if localInterestsHash != DeviceStateStore.interestsService.getServerConfirmedInterestsHash() {
+        if localInterestsHash != DeviceStateStore().getServerConfirmedInterestsHash() {
             let result = self.networkService.setSubscriptions(instanceId: self.instanceId, deviceId: Device.getDeviceId()!, interests: localInterests, retryStrategy: JustDont())
             if case .value(()) = result {
-                DeviceStateStore.interestsService.persistServerConfirmedInterestsHash(localInterestsHash)
+                DeviceStateStore().persistServerConfirmedInterestsHash(localInterestsHash)
             }
         }
     }
@@ -202,14 +202,14 @@ class ServerSyncProcessHandler {
             let localIntersets: [String] = DeviceStateStore.synchronize {
                 Device.persist(device.id)
                 Device.persistAPNsToken(token: token)
-                return DeviceStateStore.interestsService.getSubscriptions() ?? []
+                return DeviceStateStore().getSubscriptions() ?? []
             }
 
             if !localIntersets.isEmpty {
                 _ = self.networkService.setSubscriptions(instanceId: self.instanceId, deviceId: device.id, interests: localIntersets, retryStrategy: WithInfiniteExpBackoff())
             }
 
-            if let userId = DeviceStateStore.usersService.getUserId() {
+            if let userId = DeviceStateStore().getUserId() {
                 let tokenProvider = self.getTokenProvider()
                 if tokenProvider == nil {
                     // Any failures during this process are equivalent to de-authing the user e.g. setUserId(null)
@@ -217,14 +217,14 @@ class ServerSyncProcessHandler {
                     // If the user session is still valid, there should be a setUserId in the backlog
 
                     print("[PushNotifications]: Warning - Failed to set the user id due token provider not being present")
-                    DeviceStateStore.usersService.removeUserId()
+                    DeviceStateStore().removeUserId()
                 } else {
                     let semaphore = DispatchSemaphore(value: 0)
                     do {
                         try tokenProvider!.fetchToken(userId: userId, completionHandler: { jwt, error in
                             if error != nil {
                                 print("[PushNotifications]: Warning - Unexpected customer error: \(error!.localizedDescription)")
-                                DeviceStateStore.usersService.removeUserId()
+                                DeviceStateStore().removeUserId()
                                 semaphore.signal()
                                 return
                             }
@@ -233,10 +233,10 @@ class ServerSyncProcessHandler {
 
                             switch result {
                             case .value:
-                                DeviceStateStore.usersService.setUserId(userId: userId)
+                                DeviceStateStore().setUserId(userId: userId)
                             case .error(let error):
                                 print("[PushNotifications]: Warning - Unexpected error: \(error.getErrorMessage())")
-                                DeviceStateStore.usersService.removeUserId()
+                                DeviceStateStore().removeUserId()
                                 semaphore.signal()
                                 return
                             }
@@ -246,7 +246,7 @@ class ServerSyncProcessHandler {
                         semaphore.wait()
                     } catch (let error) {
                         print("[PushNotifications]: Warning - Unexpected error: \(error.localizedDescription)")
-                        DeviceStateStore.usersService.removeUserId()
+                        DeviceStateStore().removeUserId()
                     }
                 }
             }
@@ -276,7 +276,7 @@ class ServerSyncProcessHandler {
 
                 switch result {
                 case .value:
-                    DeviceStateStore.usersService.setUserId(userId: userId)
+                    DeviceStateStore().setUserId(userId: userId)
                     self.handleServerSyncEvent(.UserIdSetEvent(userId: userId, error: nil))
                 case .error(let error):
                     let error = TokenProviderError.error("[PushNotifications] - Error when synchronising with server: \(error)")
