+++ Sources/Persistence/PersistenceConstants.swift
@@ -15,6 +15,7 @@ struct PersistenceConstants {
         static let prefix = "com.pusher.sdk.interests"
         static let userId = "com.pusher.sdk.user.id"
         static let hashKey = "interestsHash"
+        static let globalScopeId = "com.pusher.sdk"
     }
 
     struct PushNotificationsInstancePersistence {
